requirements.txt

boto3==1.13.11
dnspython==1.16.0
libvirt-python==4.5.0
pyyaml==5.3.1
kubernetes==11.0.0
requests==2.24.0
retry==0.9.2
tqdm==4.46.0
waiting==1.4.1
netaddr==0.8.0
dataclasses

# Test
pre-commit==2.5.1
hypothesis==5.35.4
pytest==5.4.3
paramiko==2.7.1